Initial `pbench_copy_data` changes from review
We ensure double quotes around `--user` and `--prefix` options to
`pbench-copy-results` to ensure empty values don't get lost and the
following text is inadvertently used.

The `usage` function no longer calls exit, so that the calling location
can exit with the proper error code, and also direct the usage output
to `stderr` or `stdout` as appropriate.

We also cleaned up tabs vs spaces.